Move point-scalar multiplication to its own module
